Fix int -> unsigned int

+++ src/flash/nor/gd32vf103.c
@@ -369,7 +369,7 @@ static int gd32vf103_protect_check(struct flash_bank *bank)
 
 		/* bit 31 controls sector 62 - 255 protection for high density
 		 * bit 31 controls sector 62 - 127 protection for connectivity line */
-		for (s = 62; s < bank->num_sectors; s++)
+		for (s = 62; s < (int)bank->num_sectors; s++)
 			bank->sectors[s].is_protected = set;
 
 		if (bank->num_sectors > 61)
@@ -400,7 +400,7 @@ static int gd32vf103_protect_check(struct flash_bank *bank)
 	return ERROR_OK;
 }
 
-static int gd32vf103_erase(struct flash_bank *bank, int first, int last)
+static int gd32vf103_erase(struct flash_bank *bank, unsigned int first, unsigned int last)
 {
 	struct target *target = bank->target;
 	int i;
@@ -440,7 +440,7 @@ static int gd32vf103_erase(struct flash_bank *bank, int first, int last)
 	if (retval != ERROR_OK)
 		return retval;
 
-	for (i = first; i <= last; i++) {
+	for (i = first; i <= (int)last; i++) {
 		retval = target_write_u32(target, gd32vf103_get_flash_reg(bank, FMC_CTL), FMC_CTL_PER);
 		if (retval != ERROR_OK)
 			return retval;
@@ -467,7 +467,7 @@ static int gd32vf103_erase(struct flash_bank *bank, int first, int last)
 	return ERROR_OK;
 }
 
-static int gd32vf103_protect(struct flash_bank *bank, int set, int first, int last)
+static int gd32vf103_protect(struct flash_bank *bank, int set, unsigned int first, unsigned int last)
 {
 	struct gd32vf103_flash_bank *gd32vf103_info = NULL;
 	struct target *target = bank->target;
@@ -527,7 +527,7 @@ static int gd32vf103_protect(struct flash_bank *bank, int set, int first, int la
 		if (last > 61)
 			last = 61;
 
-		for (i = first; i <= last; i++) {
+		for (i = first; i <= (int)last; i++) {
 			reg = (i / gd32vf103_info->ppage_size) / 8;
 			bit = (i / gd32vf103_info->ppage_size) - (reg * 8);
 
@@ -538,7 +538,7 @@ static int gd32vf103_protect(struct flash_bank *bank, int set, int first, int la
 		}
 	} else {
 		/* medium density flash */
-		for (i = first; i <= last; i++) {
+		for (i = first; i <= (int)last; i++) {
 			reg = (i / gd32vf103_info->ppage_size) / 8;
 			bit = (i / gd32vf103_info->ppage_size) - (reg * 8);
 
@@ -1255,7 +1255,7 @@ static int gd32vf103_mass_erase(struct flash_bank *bank)
 
 COMMAND_HANDLER(gd32vf103_handle_mass_erase_command)
 {
-	int i;
+	unsigned int i;
 
 	if (CMD_ARGC < 1)
 		return ERROR_COMMAND_SYNTAX_ERROR;
